WIP: Update Web Components Generation Script
